Adjusted after code review: Added private header file for CPU sampler, changed linux to cache result on constructor, unsupported platforms return null
