+++ source/cpu_usage_sampler.c
@@ -4,21 +4,7 @@
  */
 
 #include <aws/common/cpu_usage_sampler.h>
-
-/*********************************************************************************************************************
- * Base implementation
- ********************************************************************************************************************/
-
-struct aws_cpu_sampler_vtable {
-    int (*aws_get_cpu_sample_fn)(struct aws_cpu_sampler *sampler, double *output);
-    void (*aws_cpu_sampler_destroy)(struct aws_cpu_sampler *sampler);
-};
-
-struct aws_cpu_sampler {
-    const struct aws_cpu_sampler_vtable *vtable;
-    struct aws_allocator *allocator;
-    void *impl;
-};
+#include <aws/common/private/cpu_usage_sampler_private.h>
 
 /*********************************************************************************************************************
  * Public operations
